refactor(orchestration): move handleAddPlayerSuggestionClick to facade layer
Move handleAddPlayerSuggestionClick from App.jsx to orchestration layer:
- Add handleAddPlayerSuggestionClickOrchestrated() to
  memberSelectionOrchestrator.js
- Replace App.jsx implementation with thin wrapper
- Export from orchestration/index.js facade
- Add export smoke test

This is a verbatim move with no logic changes. Function body is
preserved exactly, only closure variables are passed via deps
(~114 lines of add-player validation and group management logic).

Dependencies: currentGroup, 7 setters, guardAddPlayerEarly,
guardAgainstGroupDuplicate, isPlayerAlreadyPlaying, getAvailableCourts,
getCourtData, saveCourtData, findMemberNumber, showAlertMessage,
CONSTANTS, Tennis global.
